requirements.txt
Django==5.1.5
djangorestframework==3.15.2
drf-yasg==1.21.8
mysqlclient==2.2.7
openpyxl==3.1.5